feat: 모드 설정, 시간 설정,계정 설정 완료
